+++ src/pages/start/form/[id].tsx
@@ -1,0 +1,10 @@
+import CreateForm from '@components/Onboarding/CreateForm'
+import StartLayout from './StartLayout'
+
+export default function ditForm() {
+  return (
+    <StartLayout>
+      <CreateForm />
+    </StartLayout>
+  )
+}
